pyproject.toml
[project]
name = "attendance-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.2.0",
    "google>=3.0.0",
    "google-api-python-client>=2.161.0",    # import googleapiclient
    "pandas>=2.2.3",
    "polars>=1.22.0",
    "pyarrow>=19.0.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tabulate>=0.9.0",
    "typer>=0.15.1",
    "xmltodict>=0.14.2",
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
gui = "src.attendance_monitoring.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ipython>=8.32.0",
    "ruff>=0.9.6",
]
